Fix clipping of SXE userdata to 32 bits

+++ libsxe/lib-sxe/sxe.h
@@ -76,7 +76,7 @@ typedef struct SXE {
     unsigned               send_buf_written;
     union {
        void            *   as_ptr;
-       int                 as_int;
+       intptr_t            as_int;
     }                      user_data;            /* Not used by sxe                                                       */
     int                    next_socket;          /* Socket to switch to from pipe when all data has been read and cleared */
 } SXE;
